--- src/qibo/backends/profiles.yml
@@ -1,18 +1,0 @@
-backends:
-  # simulation backends - numpy is available by default
-  - name: qibojit
-    driver: qibojit.custom_operators.JITCustomBackend
-    minimum_version: 0.0.4rc0
-
-  - name: tensorflow
-    driver: qibo.backends.tensorflow.TensorflowBackend
-    minimum_version: 2.2.0
-
-  # hardware backends
-  - name: qibolab
-    driver: qibolab.backend.QibolabBackend
-    minimum_version: 0.0.1.dev0
-    is_hardware: True
-
-# default active backend after importing all modules
-default: qibojit
